Move Update shared login to commons
