+++ app/build.gradle.kts
@@ -62,12 +62,12 @@ dependencies {
     androidTestImplementation(libs.androidx.ui.test.junit4)
     debugImplementation(libs.androidx.ui.tooling)
     debugImplementation(libs.androidx.ui.test.manifest)
-    //implementation(project(":library"))
-    //implementation(project(":helper"))
-    //implementation(project(":commons"))
-    implementation(libs.youtubedl.boom.helper)
-    implementation(libs.youtubedl.boom)
-    implementation(libs.youtubedl.boom.commons)
+    implementation(project(":library"))
+    implementation(project(":helper"))
+    implementation(project(":commons"))
+    //implementation(libs.youtubedl.boom.helper)
+    //implementation(libs.youtubedl.boom)
+    //implementation(libs.youtubedl.boom.commons)
 
     implementation(libs.timber)
     implementation(libs.coil.compose)
