+++ commons/src/main/java/com/farimarwat/commons/UpdateStatus.kt
@@ -1,0 +1,5 @@
+package com.farimarwat.commons
+
+enum class UpdateStatus {
+    DONE, ALREADY_UP_TO_DATE
+}
